Fix linking error
This line was caused CMake to generate a linker command line with the
argument `-o .` for the executable `StrongridDLLStressTest`, causing the
linker to fail with the error message
```
ld: open() failed, errno=21 for '.'
```
because `.` is a directory and hence not a valid filename for an output
executable.

Removing this line fixes the issue, with CMake now correctly generating
the argument `-o StrongridDLLStressTest` for the linking command.